+++ src/events.ts
@@ -1,0 +1,15 @@
+export class Events {
+    static BOUNDSCHANGED = `bounds-change`;
+
+    static MAPCLICK = 'map-click';
+    static MAPOVERLAYCOMPLETE = 'map-overlay-complete';
+
+    static MARKERRENDERED = 'marker-render';
+    static MARKERCLICK = 'marker-click';
+    static MARKERMOUSEOVER = 'marker-mouse-over';
+    static MARKERMOUSEOUT = 'marker-mouse-out';
+
+    static POLYGONCLICK = 'polygon-click';
+
+    static INFOWINDOWSHOW = 'info-window-show';
+}
